pyproject.toml
[project]
name = "mlx-textgen"
description = "An OpenAI-compatible API LLM engine with smart prompt caching, batch processing, structured output with guided decoding, and function calling for all models using MLX."
dynamic = ["version"]

authors = [
    {name = "Nathan Tam", email = "nathan1295@gmail.com"},
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS :: MacOS X",
]
dependencies = [
    "mlx-lm>=0.24.0",
    "mlx-vlm>=0.1.26",
    "outlines>=0.2.3",
    "fastapi",
    "uvicorn"
]

[project.urls]
Homepage = "https://github.com/nath1295/MLX-Textgen"

[project.scripts]
mlx_textgen = "mlx_textgen.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.setuptools.dynamic]
version = {attr = "mlx_textgen.__version__"}

[tool.setuptools.packages.find]
where = ["src"]